🔧 Rename(#11): move file and add common-components folder
