+++ src/components/common-components/button/Button.tsx
@@ -29,7 +29,7 @@ const style: {
     default: "bg-purple-main1 border-purple-main1 text-white",
     hover: "text-white bg-purple-main2 border-purple-main2",
     border: "text-purple-main1 border-purple-main1 bg-white",
-    gray: "bg-[#F5F5F5] text-gray-60",
+    gray: "bg-[#F5F5F5] text-gray-60 border-[#F5F5F5]",
     black: "bg-black text-white",
     disabled: "bg-gray-40 text-white border-gray-40",
   },
